refactor(scripts): split PR review status
