+++ scripts/pr-review-status.ts
@@ -2,753 +2,42 @@
 
 import { resolve } from 'node:path';
 import { fileURLToPath } from 'node:url';
-import {
-  ghGraphql,
-  ghJson,
-  isCheckEntries,
-  isPrNumberPayload,
-  isPullRequestGraphqlResponse,
-  isReviewThreadsResponse,
-} from './pr-review-status-parse.js';
-
-interface PullRequestView {
-  readonly number: number;
-  readonly title: string;
-  readonly state: string;
-  readonly baseRefName: string;
-  readonly headRefName: string;
-  readonly isDraft: boolean;
-  readonly url: string;
-  readonly reviewDecision?: string | null;
-  readonly mergeStateStatus?: string | null;
-  readonly comments: readonly PullRequestComment[];
-  readonly reviews: readonly PullRequestReview[];
-}
-
-export interface CheckEntry {
-  readonly name: string;
-  readonly bucket: string;
-  readonly state: string;
-  readonly link?: string;
-  readonly workflow?: string;
-}
-
-export interface PullRequestComment {
-  readonly author?: { readonly login: string; readonly __typename?: string } | null;
-  readonly body: string;
-  readonly createdAt: string;
-}
-
-export interface PullRequestReview {
-  readonly author?: { readonly login: string; readonly __typename?: string } | null;
-  readonly body: string;
-  readonly submittedAt: string | null;
-  readonly state: string;
-}
-
-export interface ReviewThreadComment {
-  readonly author?: { readonly login: string } | null;
-  readonly body: string;
-  readonly path?: string | null;
-  readonly url: string;
-}
-
-export interface ReviewThreadNode {
-  readonly isResolved: boolean;
-  readonly comments: { readonly nodes: readonly ReviewThreadComment[] };
-}
-
-export interface ReviewThreadsResponse {
-  readonly data: {
-    readonly repository: {
-      readonly pullRequest: {
-        readonly reviewThreads: PullRequestConnection<ReviewThreadNode>;
-      };
-    };
-  };
-}
-
-export interface PageInfo {
-  readonly hasNextPage: boolean;
-  readonly endCursor?: string | null;
-}
-
-export interface PullRequestConnection<TNode> {
-  readonly nodes: readonly TNode[];
-  readonly totalCount: number;
-  readonly pageInfo: PageInfo;
-}
-
-interface PullRequestGraphqlNode extends Omit<PullRequestView, 'comments' | 'reviews'> {
-  readonly comments: PullRequestConnection<PullRequestComment>;
-  readonly reviews: PullRequestConnection<PullRequestReview>;
-}
-
-export interface PullRequestGraphqlResponse {
-  readonly data: {
-    readonly repository: {
-      readonly pullRequest: PullRequestGraphqlNode | null;
-    };
-  };
-}
-
-export interface CheckSummary {
-  readonly counts: Readonly<Record<string, number>>;
-  readonly failing: readonly CheckEntry[];
-  readonly pending: readonly CheckEntry[];
-  readonly canceled: readonly CheckEntry[];
-  readonly codeRabbit: CheckEntry | null;
-}
-
-export interface UnresolvedFinding {
-  readonly author: string;
-  readonly path: string;
-  readonly url: string;
-  readonly summary: string;
-}
-
-export interface ReviewSummary {
-  readonly total: number;
-  readonly approvals: number;
-  readonly changesRequested: number;
-  readonly comments: number;
-  readonly byState: Readonly<Record<string, number>>;
-}
-type CodeRabbitEventKind = 'rate_limit' | 'clean' | 'actionable' | 'other';
-type SubmittedReview = PullRequestReview & { readonly submittedAt: string };
-const S = String;
-const staleNote = (count: number): string => `${S(count)} stale rate-limit comment${count === 1 ? '' : 's'} ignored`;
-export interface CodeRabbitStatus {
-  readonly state: 'missing' | 'pass' | 'pending' | 'failing' | 'rate_limited' | 'actionable' | 'clean' | 'commented';
-  readonly detail: string;
-  readonly latestKind: CodeRabbitEventKind | 'none';
-  readonly staleRateLimitCount: number;
-  readonly activeRateLimitCount: number;
-}
-
-export interface MergeReadiness {
-  readonly status: 'ready' | 'pending' | 'blocked';
-  readonly reasons: readonly string[];
-}
-
-export function mergeReadinessHeading(readiness: MergeReadiness): string {
-  return readiness.status === 'blocked' ? 'Merge blockers' : 'Pending merge signals';
-}
-
-export function assertUntruncatedPullRequestData(input: {
-  readonly comments?: Pick<PullRequestConnection<PullRequestComment>, 'pageInfo' | 'totalCount'>;
-  readonly reviews?: Pick<PullRequestConnection<PullRequestReview>, 'pageInfo' | 'totalCount'>;
-  readonly reviewThreads?: Pick<PullRequestConnection<ReviewThreadNode>, 'pageInfo' | 'totalCount'>;
-}): void {
-  const truncated: string[] = [];
-
-  if (input.comments?.pageInfo.hasNextPage) {
-    truncated.push(`comments=${S(input.comments.totalCount)}`);
-  }
-
-  if (input.reviews?.pageInfo.hasNextPage) {
-    truncated.push(`reviews=${S(input.reviews.totalCount)}`);
-  }
-
-  if (input.reviewThreads?.pageInfo.hasNextPage) {
-    truncated.push(`reviewThreads=${S(input.reviewThreads.totalCount)}`);
-  }
-
-  if (truncated.length > 0) {
-    throw new Error(`pull request metadata truncated; pagination required for ${truncated.join(', ')}`);
-  }
-}
-
-export function summarizeChecks(checks: readonly CheckEntry[]): CheckSummary {
-  const counts: Record<string, number> = {
-    pass: 0,
-    fail: 0,
-    pending: 0,
-    skipping: 0,
-    cancel: 0,
-    other: 0,
-  };
-
-  for (const check of checks) {
-    if (check.bucket in counts) {
-      counts[check.bucket] = (counts[check.bucket] ?? 0) + 1;
-    } else {
-      counts.other = (counts.other ?? 0) + 1;
-    }
-  }
-
-  return {
-    counts,
-    failing: checks.filter((check) => check.bucket === 'fail'),
-    pending: checks.filter((check) => check.bucket === 'pending'),
-    canceled: checks.filter((check) => check.bucket === 'cancel'),
-    codeRabbit: checks.find((check) => check.name === 'CodeRabbit') ?? null,
-  };
-}
-
-export function extractUnresolvedFindings(threads: readonly ReviewThreadNode[]): readonly UnresolvedFinding[] {
-  return threads
-    .filter((thread) => !thread.isResolved)
-    .map((thread) => thread.comments.nodes[0])
-    .filter((comment): comment is ReviewThreadComment => comment != null)
-    .map((comment) => ({
-      author: comment.author?.login ?? '(unknown)',
-      path: comment.path ?? '(no file)',
-      url: comment.url,
-      summary: summarizeComment(comment.body),
-    }));
-}
-
-export function summarizeReviews(reviews: readonly PullRequestReview[]): ReviewSummary {
-  const latestByReviewer = new Map<string, SubmittedReview>();
-
-  const byState: Record<string, number> = {};
-
-  for (const review of reviews) {
-    if (!isSubmittedReview(review)) {
-      continue;
-    }
-
-    const reviewer = review.author?.login ?? '(unknown)';
-    if (isAutomatedReviewer(review.author)) {
-      continue;
-    }
-
-    const previous = latestByReviewer.get(reviewer);
-    if (previous == null || previous.submittedAt < review.submittedAt) {
-      latestByReviewer.set(reviewer, review);
-    }
-  }
-
-  for (const review of latestByReviewer.values()) {
-    byState[review.state] = (byState[review.state] ?? 0) + 1;
-  }
-
-  return {
-    total: latestByReviewer.size,
-    approvals: byState.APPROVED ?? 0,
-    changesRequested: byState.CHANGES_REQUESTED ?? 0,
-    comments: byState.COMMENTED ?? 0,
-    byState,
-  };
-}
-
-export function summarizeCodeRabbitStatus(
-  codeRabbitCheck: CheckEntry | null,
-  comments: readonly PullRequestComment[],
-  reviews: readonly PullRequestReview[],
-): CodeRabbitStatus {
-  const events = collectCodeRabbitEvents(comments, reviews);
-  const latestEvent = events[0] ?? null;
-  const latestNonRateLimitAt = events.find((event) => event.kind !== 'rate_limit')?.at ?? null;
-  const hasPassingCheck = codeRabbitCheck?.bucket === 'pass';
-
-  const staleRateLimitCount = events.filter((event) => {
-    if (event.kind !== 'rate_limit') {
-      return false;
-    }
-
-    if (hasPassingCheck) {
-      return true;
-    }
-
-    return latestNonRateLimitAt != null && event.at < latestNonRateLimitAt;
-  }).length;
-
-  const activeRateLimitCount = events.filter((event) => event.kind === 'rate_limit').length - staleRateLimitCount;
-
-  if (codeRabbitCheck?.bucket === 'pass') {
-    return {
-      state: 'pass',
-      detail: staleRateLimitCount > 0
-        ? `pass (${staleNote(staleRateLimitCount)})`
-        : 'pass',
-      latestKind: latestEvent?.kind ?? 'none',
-      staleRateLimitCount,
-      activeRateLimitCount,
-    };
-  }
-
-  if (codeRabbitCheck?.bucket === 'pending') {
-    return {
-      state: activeRateLimitCount > 0 ? 'rate_limited' : 'pending',
-      detail: activeRateLimitCount > 0 ? 'rate-limited' : 'pending',
-      latestKind: latestEvent?.kind ?? 'none',
-      staleRateLimitCount,
-      activeRateLimitCount,
-    };
-  }
-
-  if (codeRabbitCheck?.bucket === 'fail' || codeRabbitCheck?.bucket === 'cancel') {
-    return {
-      state: 'failing',
-      detail: codeRabbitCheck.bucket === 'cancel' ? 'canceled' : 'failing',
-      latestKind: latestEvent?.kind ?? 'none',
-      staleRateLimitCount,
-      activeRateLimitCount,
-    };
-  }
-
-  if (activeRateLimitCount > 0) {
-    return {
-      state: 'rate_limited',
-      detail: 'rate-limited',
-      latestKind: latestEvent?.kind ?? 'none',
-      staleRateLimitCount,
-      activeRateLimitCount,
-    };
-  }
-
-  if (latestEvent?.kind === 'clean') {
-    return {
-      state: 'clean',
-      detail: staleRateLimitCount > 0
-        ? `no actionable comments (${staleNote(staleRateLimitCount)})`
-        : 'no actionable comments',
-      latestKind: latestEvent.kind,
-      staleRateLimitCount,
-      activeRateLimitCount,
-    };
-  }
-
-  if (latestEvent?.kind === 'actionable') {
-    return {
-      state: 'actionable',
-      detail: staleRateLimitCount > 0
-        ? `actionable comments (${staleNote(staleRateLimitCount)})`
-        : 'actionable comments',
-      latestKind: latestEvent.kind,
-      staleRateLimitCount,
-      activeRateLimitCount,
-    };
-  }
-
-  if (latestEvent != null) {
-    return {
-      state: 'commented',
-      detail: staleRateLimitCount > 0
-        ? `commented (${staleNote(staleRateLimitCount)})`
-        : 'commented',
-      latestKind: latestEvent.kind,
-      staleRateLimitCount,
-      activeRateLimitCount,
-    };
-  }
-
-  return {
-    state: 'missing',
-    detail: 'missing',
-    latestKind: 'none',
-    staleRateLimitCount,
-    activeRateLimitCount,
-  };
-}
-
-export function computeMergeReadiness(input: {
-  readonly pr: Pick<PullRequestView, 'state' | 'isDraft' | 'reviewDecision' | 'mergeStateStatus'>;
-  readonly checks: CheckSummary;
-  readonly unresolvedCount: number;
-  readonly reviews: ReviewSummary;
-  readonly codeRabbit: CodeRabbitStatus;
-  readonly minReviews: number;
-}): MergeReadiness {
-  const reasons: string[] = [];
-
-  if (input.pr.state !== 'OPEN') {
-    reasons.push(`pull request is ${input.pr.state.toLowerCase()}`);
-  }
-
-  if (input.pr.isDraft) {
-    reasons.push('pull request is still a draft');
-  }
-
-  if (input.pr.reviewDecision === 'REVIEW_REQUIRED') {
-    reasons.push('review decision is review_required');
-  }
-
-  if (input.pr.reviewDecision === 'CHANGES_REQUESTED') {
-    reasons.push('review decision is changes_requested');
-  }
-
-  if (input.pr.mergeStateStatus != null && input.pr.mergeStateStatus !== '' && isBlockingMergeState(input.pr.mergeStateStatus)) {
-    reasons.push(`merge state is ${input.pr.mergeStateStatus.toLowerCase()}`);
-  }
-
-  if (input.checks.failing.length > 0) {
-    reasons.push(`${S(input.checks.failing.length)} failing check${input.checks.failing.length === 1 ? '' : 's'}`);
-  }
-
-  if (input.checks.canceled.length > 0) {
-    reasons.push(`${S(input.checks.canceled.length)} canceled check${input.checks.canceled.length === 1 ? '' : 's'}`);
-  }
-
-  if (input.unresolvedCount > 0) {
-    reasons.push(`${S(input.unresolvedCount)} unresolved review thread${input.unresolvedCount === 1 ? '' : 's'}`);
-  }
-
-  if (input.reviews.changesRequested > 0) {
-    reasons.push(`${S(input.reviews.changesRequested)} change-request review${input.reviews.changesRequested === 1 ? '' : 's'}`);
-  }
-
-  if (input.reviews.total < input.minReviews) {
-    reasons.push(`needs at least ${S(input.minReviews)} review${input.minReviews === 1 ? '' : 's'} (found ${S(input.reviews.total)})`);
-  }
-
-  if (reasons.length > 0) {
-    return { status: 'blocked', reasons };
-  }
-
-  const pendingReasons: string[] = [];
-
-  if (input.pr.mergeStateStatus === 'UNKNOWN') {
-    pendingReasons.push('mergeability is still being computed');
-  }
-
-  if (input.checks.pending.length > 0) {
-    pendingReasons.push(`${S(input.checks.pending.length)} pending check${input.checks.pending.length === 1 ? '' : 's'}`);
-  }
-
-  if (input.codeRabbit.state === 'pending') {
-    pendingReasons.push('CodeRabbit review still pending');
-  } else if (input.codeRabbit.state === 'rate_limited') {
-    pendingReasons.push('CodeRabbit is rate-limited');
-  }
-
-  if (pendingReasons.length > 0) {
-    return { status: 'pending', reasons: pendingReasons };
-  }
-
-  return { status: 'ready', reasons: [] };
-}
-
-export function computeExitCode(summary: CheckSummary, unresolvedCount: number): number {
-  if (summary.failing.length > 0 || summary.canceled.length > 0 || unresolvedCount > 0) {
-    return 1;
-  }
-
-  if (summary.pending.length > 0) {
-    return 8;
-  }
-
-  return 0;
-}
-
-export function computeMergeReadinessExitCode(readiness: MergeReadiness): number {
-  switch (readiness.status) {
-    case 'ready':
-      return 0;
-    case 'pending':
-      return 8;
-    case 'blocked':
-      return 1;
-  }
-}
-
-function main(): void {
-  const options = parseArgs(process.argv.slice(2));
-  const pr = fetchPullRequest(options.prArg);
-  const checks = ghJson([
-    'pr',
-    'checks',
-    S(pr.number),
-    '--json',
-    'name,bucket,state,link,workflow',
-  ], isCheckEntries);
-  const threads = fetchReviewThreads(pr.number);
-  const unresolved = extractUnresolvedFindings(threads);
-  const summary = summarizeChecks(checks);
-  const reviewSummary = summarizeReviews(pr.reviews);
-  const codeRabbitStatus = summarizeCodeRabbitStatus(summary.codeRabbit, pr.comments, pr.reviews);
-  const mergeReadiness = computeMergeReadiness({
-    pr,
-    checks: summary,
-    unresolvedCount: unresolved.length,
-    reviews: reviewSummary,
-    codeRabbit: codeRabbitStatus,
-    minReviews: options.minReviews,
-  });
-
-  process.stdout.write(`PR #${S(pr.number)}: ${pr.title}\n`);
-  process.stdout.write(`State: ${pr.state}${pr.isDraft ? ' (draft)' : ''}\n`);
-  process.stdout.write(`Branch: ${pr.headRefName} -> ${pr.baseRefName}\n`);
-  if (pr.reviewDecision) {
-    process.stdout.write(`Review decision: ${pr.reviewDecision}\n`);
-  }
-  if (pr.mergeStateStatus) {
-    process.stdout.write(`Merge state: ${pr.mergeStateStatus}\n`);
-  }
-  process.stdout.write(`URL: ${pr.url}\n\n`);
-
-  process.stdout.write(
-    `Checks: pass=${S(summary.counts.pass ?? 0)} pending=${S(summary.counts.pending ?? 0)} fail=${S(summary.counts.fail ?? 0)} skipping=${S(summary.counts.skipping ?? 0)} cancel=${S(summary.counts.cancel ?? 0)}\n`,
-  );
-  process.stdout.write(`Unresolved threads: ${S(unresolved.length)}\n`);
-  process.stdout.write(
-    `Reviews: total=${S(reviewSummary.total)} approvals=${S(reviewSummary.approvals)} changes_requested=${S(reviewSummary.changesRequested)} comments=${S(reviewSummary.comments)}\n`,
-  );
-  process.stdout.write(`CodeRabbit: ${codeRabbitStatus.detail}\n`);
-
-  if (options.mergeReady) {
-    process.stdout.write(`Merge readiness: ${mergeReadiness.status.toUpperCase()}\n`);
-    process.stdout.write(`Review gate: ${S(reviewSummary.total)}/${S(options.minReviews)}\n`);
-  }
-
-  if (summary.failing.length > 0) {
-    process.stdout.write('\nFailing checks:\n');
-    for (const check of summary.failing) {
-      process.stdout.write(`- ${check.name} (${check.state})${check.link ? ` ${check.link}` : ''}\n`);
-    }
-  }
-
-  if (summary.canceled.length > 0) {
-    process.stdout.write('\nCanceled checks:\n');
-    for (const check of summary.canceled) {
-      process.stdout.write(`- ${check.name} (${check.state})${check.link ? ` ${check.link}` : ''}\n`);
-    }
-  }
-
-  if (summary.pending.length > 0) {
-    process.stdout.write('\nPending checks:\n');
-    for (const check of summary.pending) {
-      process.stdout.write(`- ${check.name} (${check.state})${check.link ? ` ${check.link}` : ''}\n`);
-    }
-  }
-
-  if (unresolved.length > 0) {
-    process.stdout.write('\nUnresolved review threads:\n');
-    for (const finding of unresolved) {
-      process.stdout.write(`- [${finding.author}] ${finding.path}: ${finding.summary}\n  ${finding.url}\n`);
-    }
-  }
-
-  if (options.mergeReady && mergeReadiness.reasons.length > 0) {
-    process.stdout.write(`\n${mergeReadinessHeading(mergeReadiness)}:\n`);
-    for (const reason of mergeReadiness.reasons) {
-      process.stdout.write(`- ${reason}\n`);
-    }
-  }
-
-  process.exitCode = options.mergeReady
-    ? computeMergeReadinessExitCode(mergeReadiness)
-    : computeExitCode(summary, unresolved.length);
-}
-
-function fetchReviewThreads(prNumber: number): readonly ReviewThreadNode[] {
-  const query = `
-query($prNumber: Int!) {
-  repository(owner: "flyingrobots", name: "bijou") {
-    pullRequest(number: $prNumber) {
-      reviewThreads(first: 100) {
-        totalCount
-        pageInfo {
-          hasNextPage
-          endCursor
-        }
-        nodes {
-          isResolved
-          comments(first: 20) {
-            nodes {
-              author { login }
-              body
-              path
-              url
-            }
-          }
-        }
-      }
-    }
-  }
-}
-`;
-
-  const payload = ghGraphql(query, { prNumber: S(prNumber) }, isReviewThreadsResponse);
-  const reviewThreads = payload.data.repository.pullRequest.reviewThreads;
-  assertUntruncatedPullRequestData({ reviewThreads });
-  return reviewThreads.nodes;
-}
-
-function summarizeComment(body: string): string {
-  const firstLine = body
-    .split('\n')
-    .map((line) => line.trim())
-    .find((line) => line.length > 0 && !line.startsWith('<!--') && !line.startsWith('<details>') && !line.startsWith('</details>'));
-
-  return firstLine == null ? '(no summary)' : firstLine.slice(0, 140);
-}
-
-function collectCodeRabbitEvents(
-  comments: readonly PullRequestComment[],
-  reviews: readonly PullRequestReview[],
-): readonly { readonly at: string; readonly kind: CodeRabbitEventKind }[] {
-  const commentEvents = comments
-    .filter((comment) => isCodeRabbitAuthor(comment.author?.login))
-    .map((comment) => ({
-      at: comment.createdAt,
-      kind: classifyCodeRabbitBody(comment.body),
-    }));
-  const reviewEvents = reviews
-    .filter(isSubmittedReview)
-    .filter((review) => isCodeRabbitAuthor(review.author?.login))
-    .map((review) => ({
-      at: review.submittedAt,
-      kind: classifyCodeRabbitBody(review.body),
-    }));
-
-  return [...commentEvents, ...reviewEvents].sort((left, right) => right.at.localeCompare(left.at));
-}
-
-function classifyCodeRabbitBody(body: string): CodeRabbitEventKind {
-  const normalized = body.toLowerCase();
-
-  if (normalized.includes('rate limit exceeded')) {
-    return 'rate_limit';
-  }
-
-  if (normalized.includes('no actionable comments were generated in the recent review')) {
-    return 'clean';
-  }
-
-  if (/\*\*actionable comments posted:\s*[1-9]/i.test(body)) {
-    return 'actionable';
-  }
-
-  if (/\*\*actionable comments posted:\s*0/i.test(body)) {
-    return 'clean';
-  }
-
-  return 'other';
-}
-
-function isAutomatedReviewer(author: PullRequestReview['author'] | PullRequestComment['author']): boolean {
-  if (author?.__typename != null) {
-    return author.__typename !== 'User';
-  }
-
-  const login = author?.login;
-  return login != null && (isCodeRabbitAuthor(login) || login.endsWith('[bot]'));
-}
-
-function isSubmittedReview(review: PullRequestReview): review is SubmittedReview {
-  return review.submittedAt != null && review.state !== 'PENDING';
-}
-
-function isBlockingMergeState(state: string): boolean {
-  return state !== 'CLEAN' && state !== 'HAS_HOOKS' && state !== 'UNSTABLE' && state !== 'UNKNOWN';
-}
-
-function isCodeRabbitAuthor(login: string | undefined): boolean {
-  if (login == null) {
-    return false;
-  }
-
-  return /^coderabbitai(?:\[bot\])?$/i.test(login);
-}
-
-function fetchPullRequest(prArg: string | undefined): PullRequestView {
-  const selector = prArg == null || prArg === '' ? undefined : prArg;
-  const prNumber = ghJson([
-    'pr',
-    'view',
-    ...maybeArg(selector),
-    '--json',
-    'number',
-  ], isPrNumberPayload).number;
-
-  const query = `
-query($prNumber: Int!) {
-  repository(owner: "flyingrobots", name: "bijou") {
-    pullRequest(number: $prNumber) {
-      number
-      title
-      state
-      baseRefName
-      headRefName
-      isDraft
-      url
-      reviewDecision
-      mergeStateStatus
-      comments(first: 100) {
-        totalCount
-        pageInfo {
-          hasNextPage
-          endCursor
-        }
-        nodes {
-          body
-          createdAt
-          author { login __typename }
-        }
-      }
-      reviews(first: 100) {
-        totalCount
-        pageInfo {
-          hasNextPage
-          endCursor
-        }
-        nodes {
-          body
-          submittedAt
-          state
-          author { login __typename }
-        }
-      }
-    }
-  }
-}
-`;
-  const payload = ghGraphql(query, { prNumber: S(prNumber) }, isPullRequestGraphqlResponse);
-  const pullRequest = payload.data.repository.pullRequest;
-  if (pullRequest == null) {
-    throw new Error(`no pull request found for ${selector ?? 'current branch'}`);
-  }
-
-  assertUntruncatedPullRequestData(pullRequest);
-
-  return {
-    ...pullRequest,
-    comments: pullRequest.comments.nodes,
-    reviews: pullRequest.reviews.nodes,
-  };
-}
-
-function maybeArg(value: string | undefined): readonly string[] {
-  return value == null || value === '' ? [] : [value];
-}
-
-function parseArgs(args: readonly string[]): { readonly mergeReady: boolean; readonly minReviews: number; readonly prArg?: string } {
-  let mergeReady = false;
-  let minReviews = 2;
-  let prArg: string | undefined;
-
-  for (let index = 0; index < args.length; index += 1) {
-    const arg = args[index] ?? '';
-    if (arg === '--merge-ready') {
-      mergeReady = true;
-      continue;
-    }
-
-    if (arg === '--min-reviews') {
-      const value = args[index + 1];
-      if (value == null || !/^\d+$/.test(value)) {
-        throw new Error('--min-reviews requires a numeric value');
-      }
-
-      minReviews = Number.parseInt(value, 10);
-      index += 1;
-      continue;
-    }
-
-    if (prArg == null) {
-      prArg = arg;
-      continue;
-    }
-
-    throw new Error(`unexpected argument: ${arg}`);
-  }
-
-  return { mergeReady, minReviews, prArg };
-}
-
-if (process.argv[1] != null && fileURLToPath(import.meta.url) === resolve(process.argv[1])) {
-  main();
+import { runReviewStatus } from './pr-review-status-cli.js';
+
+export type {
+  CheckEntry,
+  CheckSummary,
+  CodeRabbitStatus,
+  MergeReadiness,
+  PullRequestComment,
+  PullRequestReview,
+  ReviewSummary,
+  UnresolvedFinding,
+} from './pr-review-status-contract.js';
+export type {
+  PageInfo,
+  PullRequestConnection,
+  PullRequestGraphqlResponse,
+  ReviewThreadComment,
+  ReviewThreadNode,
+  ReviewThreadsResponse,
+} from './pr-review-status-github-contract.js';
+export { computeExitCode, summarizeChecks } from './pr-review-status-checks.js';
+export { summarizeCodeRabbitStatus } from './pr-review-status-coderabbit.js';
+export {
+  computeMergeReadiness,
+  computeMergeReadinessExitCode,
+  mergeReadinessHeading,
+} from './pr-review-status-readiness.js';
+export {
+  assertUntruncatedPullRequestData,
+  extractUnresolvedFindings,
+  summarizeReviews,
+} from './pr-review-status-reviews.js';
+
+if (
+  process.argv[1] != null &&
+  fileURLToPath(import.meta.url) === resolve(process.argv[1])
+) {
+  runReviewStatus(process.argv.slice(2));
 }
